pyproject.toml
[project]
name = "jflow"
version = "0.1.0"
description = "Renders Jinja2 templates from a declarative YAML config"
readme = "README.md"
authors = [
    { name = "Ovidiu Gheorghies", email = "ogheorghies@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "jinja2",
    "openpyxl",
    "pyyaml",
]

[project.scripts]
jflow = "jflow:main"

[build-system]
requires = ["uv_build>=0.9.26,<0.10.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
